add puck info to config fix pin numbers in puck keep list of pucks in puck scanner_data is now a field in the scanner module add some methods in db to get objects from db add puck_address as field in LiquorInventory class setup main to init all code, run update loop, shutdown add beginnings of command interpreter

+++ src/scorpion/localdb/db.py
@@ -9,25 +9,38 @@ import sqlalchemy as sql
 import sqlalchemy.orm as orm
 
 import scorpion.config as config
-import dbobjects
+import dbobjects as dbo
 import xmlparser
 
+session = None
+
+def get_inventory():
+    inventory_liquor = session.query(dbo.LiquorInventory).all()
+    inventory_extra = session.query(dbo.ExtraInventory).all()
+    return (inventory_liquor, inventory_extra)
+
+def get_drinks():
+    drinks = session.query(dbo.Drink).all()
+    return drinks
+
+def get_drink(name):
+    drink = session.query(dbo.Drink).filter(dbo.Drink.name == name).first()
+    return drink
 
 def put_some_data(session):
     objects = xmlparser.get_objects()
     session.add_all(objects)
 
-def reset_db():
-    if os.path.exists(config.local_db):
+def init_db(reset = False):
+    global session
+    if reset and os.path.exists(config.local_db):
         os.remove(config.local_db)
         print 'removed old db'
-    
     engine = sql.create_engine('sqlite:///'+config.local_db, echo = True)
-    dbobjects.create_tables(engine)
+    dbo.create_tables(engine)
     Session = orm.sessionmaker(bind=engine)
     session = Session()
-    put_some_data(session)
-    session.commit()
-    
-if __name__ == "__main__":
-    reset_db()
+    if reset: 
+        put_some_data(session)
+        session.commit()
+
